Some substantial changes to argument processing.
Argument processing is now modular and configurable. Different aspects
of populating the args and kwargs passed to the endpoints are now split
among several extension classes which cooperate in mutating the
arguments.

This netted a 75 r/sec (5%) improvement in performance, slimmed down the
Application class, as well as benefitting from better handling of rich
data stored in flat namespaces.